Refactore session state to sub-package
